Added Gameplay Assets
Player was added to the 1.1 scene, The item pickup script was added to the black cube. This puts a test key item in the players inventory. The lock script was put on the red box. This searches the player's inventory for the key item. It doesn't actually do anything while unlocked though. Both of these need triggers for them to work.